chore: Rename fastify example, show body parsing, and link logger
