+++ examples/fastify/package.json
@@ -1,4 +1,5 @@
 {
+  "name": "examples-fastify",
   "dependencies": {
     "@arcjet/fastify": "file:../../arcjet-fastify",
     "fastify": "^5.0.0"
